feat: Build both `amd64` and `arm64` docker images.
Prior to this commit, the built docker image was only for `amd64`,
and users on other platforms (notably the Mac M1) could only run it
under emulation. That's basically a non-starter for CPU-intensive
tasks like compiling Rust code, and leads to intolerably slow builds
on such platforms.

This commit re-works the Dockerfile and surrounding build tooling
to be able to publish both `amd64` and `arm64` images. Preliminary
tests show that the `arm64` image can produce much faster build
times on aarch64 platforms like the Mac M1.